feat(client): add news store and restructure api types
- Create news store and type definitions.
- Moved shared types from auth to news directory.
- Update imports in app to reflect type changes.
